monitor.js catch errors
I removed the hacky and inaccurate `if (money === 0) money = 1;` statement. This would cause `money > maxMoney`.
In place of this I added 2 statements related to `money < 1` which triggered `$_______: $0.000 / $0.000 (NaN%)` to print. 
This should be fine but doesn't make it easy at a glance that something is wrong. 
So a Error statement is printed, turning it red with an explanation.
Variables can be made static text but isn't required so I didn't bother.

Then for the `grow____:` line, it would crash because of `maxMoney / money`, example: `0 / 1` as it was before and `0/0` after removing `if (money === 0) money = 1;`
Both cases it expects a number above 1 but it isn't returned. Instead of patching it with another lied 1 as money had, I check `money < 1` and print an error.

This results in the following:
![New result without script crash.](https://i.imgur.com/LqOrlf4.pngg)

Instead of:
![Original result that created an error causing a script crash.](https://i.imgur.com/AsxvKY6.png)